# openapi/authz-api.yaml
openapi: 3.0.3
info:
  title: Busibox AuthZ API
  description: |
    Authorization and authentication service for Busibox platform.
    
    Provides:
    - OAuth2 token issuance (client_credentials and token-exchange grants)
    - RBAC management (roles, users, permissions)
    - Audit logging
    - JWKS endpoint for token validation
    
    **Authentication**: OAuth client credentials or admin token required for most endpoints.
  version: 1.0.0
  contact:
    name: Busibox Team
servers:
  - url: http://10.96.200.210:8010
    description: Production AuthZ Service
  - url: http://10.96.201.210:8010
    description: Test AuthZ Service
...
paths:
...
  /oauth/token:
    post:
      tags: [OAuth]
      summary: OAuth2 Token Endpoint
      description: |
        Issue access tokens using OAuth2 grants.
        
        Supported grant types:
        - `client_credentials`: Service-to-service authentication
        - `urn:ietf:params:oauth:grant-type:token-exchange`: On-behalf-of user tokens (RFC 8693)
        
        Accepts both `application/x-www-form-urlencoded` and `application/json` content types.
      operationId: token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthTokenRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OAuthTokenRequest'
      responses:
        '200':
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized audience or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    AdminToken:
      type: http
      scheme: bearer
      description: Admin token for RBAC operations (AUTHZ_ADMIN_TOKEN)

    OAuthClient:
      type: oauth2
      description: OAuth client credentials flow
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes: {}
  schemas:
...
    OAuthTokenRequest:
      type: object
      required:
        - grant_type
        - client_id
        - client_secret
      properties:
        grant_type:
          type: string
          enum:
            - client_credentials
            - urn:ietf:params:oauth:grant-type:token-exchange
          description: OAuth2 grant type
        client_id:
          type: string
          description: OAuth client ID
        client_secret:
          type: string
          description: OAuth client secret
        audience:
          type: string
          description: Target service (e.g., ingest-api, search-api, agent-api)
          example: ingest-api
        scope:
          type: string
          description: Space-delimited OAuth2 scopes
          example: ingest.write ingest.read
        requested_subject:
          type: string
          format: uuid
          description: User ID for token-exchange grant
        requested_purpose:
          type: string
          description: Purpose label for audit logging
          example: document-upload
        subject_token:
          type: string
          description: RFC 8693 subject token (optional)
        subject_token_type:
          type: string
          description: RFC 8693 subject token type (optional)
    OAuthTokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 900
        scope:
          type: string
          description: Granted scopes (space-delimited)
          example: ingest.write ingest.read
        issued_token_type:
          type: string
          description: RFC 8693 issued token type
          example: urn:ietf:params:oauth:token-type:access_token
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        detail:
          type: string
          description: Error details

    # ========================================================================
    # Keystore Schemas (Envelope Encryption)
    # ========================================================================
...